@@ -1,61 +1,47 @@
 package ranges
 
 import (
-	"fmt"
-	"sort"
+	"cmp"
+	"iter"
+	"slices"
 	"strconv"
 	"strings"
 	"unsafe"
 
 	"golang.org/x/exp/constraints"
-	"golang.org/x/exp/slices"
 )
 
 type Ranges[I constraints.Integer] []Range[I]
-
-func (r Range[I]) Plural() Ranges[I] {
-	if r.Len() == 0 {
-		return nil
-	}
-	return Ranges[I]{r}
-}
 
 func (r Ranges[I]) Equal(other Ranges[I]) bool {
 	return slices.Equal(r, other)
 }
 
-func (r Ranges[I]) binsearch(x I) (int, bool) {
-	index := sort.Search(len(r), func(i int) bool { return x <= r[i].end })
-	switch {
-	case index == -1:
-		return len(r), false
-	case index >= len(r):
-		return index, false
-	default:
-		return index, r[index].has(x)
-	}
-}
-
-func (r Ranges[I]) Binsearch(x I) int {
-	if len(r) == 0 || x < r[0].start {
-		return -1
-	}
-	index := sort.Search(len(r), func(i int) bool { return x <= r[i].end })
-	switch {
-	case index == -1:
-		return index
-	case index >= len(r):
-		return -int(r.Len())
-	default:
-		offset := int(r[:index].Len())
-		if r[index].has(x) {
-			return offset + int(x-r[index].start)
+func (r Ranges[I]) Binsearch(value I) (int, bool) {
+	return slices.BinarySearchFunc(r, value, func(item Range[I], value I) int {
+		if item.has(value) {
+			return 0
 		}
-		return -offset - 1
-	}
+		return cmp.Compare(item.start, value)
+	})
 }
 
-func (r Ranges[I]) Contains(x I) bool {
+func (r Ranges[I]) Contains(other Ranges[I]) bool {
+	index := 0
+	for _, chunk := range other {
+		offset, ok := r[index:].Binsearch(chunk.start)
+		if !ok {
+			return false
+		}
+		index += offset
+		if r[index].end < chunk.end {
+			return false
+		}
+	}
+	return true
+}
+
+func (r Ranges[I]) Has(x I) bool {
 	if len(r) == 0 {
 		return false
 	}
@@ -63,8 +49,13 @@ func (r Ranges[I]) Contains(x I) bool {
 		return false
 	}
 
-	_, found := r.binsearch(x)
+	_, found := r.Binsearch(x)
 	return found
+}
+
+func (r Ranges[I]) ContainsNaive(naive Range[I]) bool {
+	index, ok := r.Binsearch(naive.start)
+	return ok && naive.end <= r[index].end
 }
 
 func (r Ranges[I]) Len() uint64 {
@@ -83,7 +74,7 @@ func (r Ranges[I]) WriteTo(buf *strings.Builder) {
 	if len(r) == 0 {
 		return
 	}
-	for i := 0; i < len(r)-1; i++ {
+	for i := range len(r) - 1 {
 		buf.WriteString(r[i].String())
 		buf.WriteRune(',')
 	}
@@ -99,96 +90,40 @@ func (r Ranges[I]) String() string {
 	return buf.String()
 }
 
-func (r Ranges[I]) Clone() Ranges[I] {
-	ranges := make(Ranges[I], len(r))
-	copy(ranges, r)
-	return ranges
-}
+func (r Ranges[I]) Clone() Ranges[I] { return slices.Clone(r) }
 
-func (r Ranges[I]) Iter(fn func(x I) bool) (I, bool) {
-	for i := 0; i < len(r); i++ {
-		for v := r[i].start; v <= r[i].end; v++ {
-			if fn(v) {
-				return v, true
+func (r Ranges[I]) Ref() *Ranges[I] { return &r }
+
+func (r *Ranges[I]) Assign() Assigner[I] { return Assigner[I]{r} }
+
+func (r Ranges[I]) Iter() iter.Seq[I] {
+	return func(yield func(I) bool) {
+		for i := range len(r) {
+			for v := r[i].start; v <= r[i].end; v++ {
+				if !yield(v) {
+					return
+				}
 			}
 		}
 	}
-	return 0, false
 }
 
 func (r Ranges[I]) NumChunks() int { return len(r) }
 
-func (r Ranges[I]) IterChunks(fn func(Range[I]) bool) Range[I] {
-	for _, chunk := range r {
-		if fn(chunk) {
-			return chunk
-		}
-	}
-	return Range[I]{1, 0}
-}
+func (r Ranges[I]) First() I { return r[0].start }
+func (r Ranges[I]) Last() I  { return r[len(r)-1].end }
 
-func (r Ranges[I]) Foreach(consumer func(x I)) {
-	for i := 0; i < len(r); i++ {
-		for v := r[i].start; v <= r[i].end; v++ {
-			consumer(v)
-		}
-	}
-}
-
-func (r Ranges[I]) First() I {
-	return r[0].start
-}
-
-func (r Ranges[I]) Last() I {
-	return r[len(r)-1].end
-}
-
-func (r *Ranges[I]) Pop() (I, bool) {
-	if len(*r) == 0 {
-		return 0, false
-	}
-	firstInterval := (*r)[0]
-	if firstInterval.Len() == 1 {
-		*r = (*r)[1:]
-	} else {
-		(*r)[0].shrinkLower(1)
-	}
-	return firstInterval.start, true
-}
+func (r Ranges[I]) firstChunk() *Range[I] { return &r[0] }
+func (r Ranges[I]) lastChunk() *Range[I]  { return &r[len(r)-1] }
 
 func (r Ranges[I]) Index(index uint64) (I, bool) {
-	for i := 0; i < len(r); i++ {
+	for i := range len(r) {
 		if index < r[i].Len() {
 			return r[i].start + I(index), true
 		}
 		index -= r[i].Len()
 	}
 	return 0, false
-}
-
-func (r Ranges[I]) Limit(lower, upper I) Ranges[I] {
-	if lower > upper {
-		panic(fmt.Sprintf("Lower %d > upper %d", lower, upper))
-	}
-	if r.IsEmpty() || lower <= r.First() && r.Last() <= upper {
-		return r
-	}
-	lowerIndex, lowerFound := r.binsearch(lower)
-	upperIndex, upperFound := r.binsearch(upper)
-	if !lowerFound && !upperFound {
-		return r[lowerIndex:upperIndex]
-	}
-	if upperFound {
-		upperIndex++
-	}
-	retval := slices.Clone(r[lowerIndex:upperIndex])
-	if lowerFound {
-		retval[0].start = lower
-	}
-	if upperFound {
-		retval[len(retval)-1].end = upper
-	}
-	return retval
 }
 
 func Cast[I, T constraints.Integer](ranges Ranges[I]) (Ranges[T], bool) {
@@ -219,7 +154,7 @@ func Cast[I, T constraints.Integer](ranges Ranges[I]) (Ranges[T], bool) {
 			maximum = I(^T(0))
 		}
 	}
-	limited := ranges.Limit(minimum, maximum)
+	limited := ranges.Limit(Range[I]{minimum, maximum})
 	retval := make(Ranges[T], len(limited))
 	for i, r := range limited {
 		retval[i] = Range[T]{start: T(r.start), end: T(r.end)}
@@ -230,7 +165,7 @@ func Cast[I, T constraints.Integer](ranges Ranges[I]) (Ranges[T], bool) {
 // Caller must ensure slice ordered, otherwise result probably invalid
 func FromSlice[I constraints.Integer](slice []I) Ranges[I] {
 	if len(slice) == 0 {
-		return Empty[I]().Plural()
+		return Empty[I]().Ranges()
 	}
 	var retval Ranges[I]
 	pending := Of(slice[0])
@@ -251,7 +186,7 @@ func FromSlice[I constraints.Integer](slice []I) Ranges[I] {
 func mergeRanges[I constraints.Integer](rangeList []Range[I]) Ranges[I] {
 	var ranges Ranges[I]
 	for _, ranging := range rangeList {
-		ranges = ranges.Add(ranging.Plural())
+		ranges = ranges.Add(ranging.Ranges())
 	}
 	return ranges
 }
