@@ -1,6 +1,10 @@
 package ranges
 
-import "golang.org/x/exp/constraints"
+import (
+	"iter"
+
+	"golang.org/x/exp/constraints"
+)
 
 type LazyRanges[I constraints.Integer] struct{ left, right Ranges[I] }
 
@@ -12,8 +16,8 @@ func (r LazyRanges[I]) Len() uint64 {
 	return r.Eval().Len()
 }
 
-func (r LazyRanges[I]) Foreach(consumer func(x I)) {
-	r.Eval().Foreach(consumer)
+func (r LazyRanges[I]) Iter() iter.Seq[I] {
+	return r.Eval().Iter()
 }
 
 func LazySub[I constraints.Integer](left, right Ranges[I]) LazyRanges[I] {
